requirements.txt
pyyaml
PyPDF2
